+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed 🐛
+
+* Fixed bug where a lost widget could still retain keyboard focus.
+
 ## 0.3.0 - 2020-11-07
 
 ### Added ⭐